refactor: rename to AnimatableHeader
