+++ src/lib/Components/AnimatableHeader/AnimatableHeaderFlatList.tsx
@@ -1,14 +1,14 @@
 import React from "react"
 import { FlatList, FlatListProps } from "react-native"
 import Animated from "react-native-reanimated"
+import { useAnimatableHeaderContext } from "./AnimatableHeaderContext"
 import { AnimatableHeaderLargeTitle } from "./AnimatableHeaderLargeTitle"
-import { useCollapsibleHeaderContext } from "./CollapsibleHeaderContext"
 
 const AnimatedComponent: typeof FlatList = Animated.createAnimatedComponent(FlatList)
 
 export const AnimatableHeaderFlatList = <T extends any>(props: FlatListProps<T>) => {
   const { ListHeaderComponent, ...other } = props
-  const { scrollOffsetY } = useCollapsibleHeaderContext()
+  const { scrollOffsetY } = useAnimatableHeaderContext()
 
   return (
     <AnimatedComponent
