+++ src/lib/Components/AnimatableHeader/AnimatableHeaderScrollView.tsx
@@ -1,14 +1,14 @@
 import React, { PropsWithChildren } from "react"
 import { FlatListProps, ScrollView } from "react-native"
 import Animated from "react-native-reanimated"
+import { useAnimatableHeaderContext } from "./AnimatableHeaderContext"
 import { AnimatableHeaderLargeTitle } from "./AnimatableHeaderLargeTitle"
-import { useCollapsibleHeaderContext } from "./CollapsibleHeaderContext"
 
 const AnimatedComponent: typeof ScrollView = Animated.createAnimatedComponent(ScrollView)
 
 export const AnimatableHeaderScrollView = <T extends any>(props: PropsWithChildren<FlatListProps<T>>) => {
   const { children, ...other } = props
-  const { scrollOffsetY } = useCollapsibleHeaderContext()
+  const { scrollOffsetY } = useAnimatableHeaderContext()
 
   return (
     <AnimatedComponent
